requirements.txt
requests>=2.31.0
pandas>=2.2.2
python-dotenv>=1.0.1
snowflake-connector-python[pandas]>=3.6.0
apache-airflow==2.9.1
